refactor trainer tests
